big improvement to install process

+++ install/symlink.sh
@@ -1,20 +1,21 @@
-rm -rf $HOME/.config/nvim
-mkdir -p $HOME/.config/nvim
+
+if [ "$1" == "fresh" ]; then
+    rm -rf $HOME/.local/bin
+    mkdir -p $HOME/.local/bin
+    rm -rf $HOME/.zfunctions
+    mkdir -p $HOME/.zfunctions
+    rm -rf $HOME/.config/nvim
+    mkdir -p $HOME/.config/nvim
+fi
+
 ln -snf $DOTFILES/nvim/* $HOME/.config/nvim
-
-rm -rf $HOME/.local/bin
-mkdir -p $HOME/.local/bin
-ln -snf $DOTFILES/scripts/* $HOME/.local/bin
-
-rm -rf $HOME/.zfunctions
-mkdir -p $HOME/.zfunctions
-ln -snf $DOTFILES/zfunctions/* $HOME/.zfunctions
-
+ln -snf $DOTFILES/tmux $HOME/.config/tmux
+ln -sf $DOTFILES/scripts/* $HOME/.local/bin
+ln -sf $DOTFILES/zfunctions/* $HOME/.zfunctions
 ln -sf $DOTFILES/.p10k.zsh $HOME/.p10k.zsh
 ln -sf $DOTFILES/.zshrc $HOME/.zshrc
 ln -sf $DOTFILES/.zprofile $HOME/.zprofile
 ln -sf $DOTFILES/.aliases $HOME/.aliases
-ln -sf $DOTFILES/tmux/.tmux.conf $HOME/.tmux.conf
-ln -sf $DOTFILES/tmux/.tmux-cht-languages $HOME/.tmux-cht-languages
-ln -sf $DOTFILES/tmux/.tmux-cht-command $HOME/.tmux-cht-command
-ln -sf $DOTFILES/.editorconfig $HOME/projects/.editorconfig
+mkdir $HOME/.config/kitty && ln -sf $DOTFILES/kitty.conf $HOME/.config/kitty/kitty.conf
+mkdir $HOME/.config/rofi && ln -sf $DOTFILES/config.rase $HOME/.config/rofi/config.rasi
+ln -sf $DOTFILE/picom.conf $HOME/.config/picom.conf
